图书管理系统/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
pymysql==1.1.0
werkzeug==2.3.7 